JMtk15 Package Documentation

Welcome to JMtk15, a Julia language package to leverage the NASA JPL MISR Toolkit (Mtk) C shared library [Version 1.5], which is a convenient and simplified programming interface to access MISR L1B2, L2, and MISR-HR, conventional and ancillary data products.

The Multi-angle Imaging SpectroRadiometer (MISR) instrument is one of five instruments aboard NASA's Earth Observation Terra satellite launched on 18 December 1999. MISR has been collecting data on the causes and effects of climate change in the atmosphere and at the planetary surface every day since 24 February 2000, and is expected to continue operating until early 2027. A wide range of geophysical products are available on a per-orbit basis at spatial resoultions varying between 275 m and 17.6 km, depending on the product. Higher-level products with a global coverage and at coarser spatial resolution are also available.

Detailed information on the scientific context of this project is available on-line from:

the NASA Terra mission,

the MISR instrument, which includes a complete bibliography of thousands of published papers and reports,

the NASA Atmospheric Science Data Center, and

the MISR Toolkit.

Package status and acknowledgments

The Julia functions described below are wrappers to a selection of some of the most useful C functions of Mtk. They are currently under active development and are expected to work as described in the associated tests and documentation, but they are provided under the MIT License, i.e., without any warranty about the their usability or effectiveness in your own applications. See the LICENSE file in this repository for more details.

Generating this package would not have been possible without the professional help of Sebastian Val at JPL concerning Mtk and the benevolent support of @Gnimuc, @stevengj as well as many other contributors with respect to the Julia implementation.

Please check this site regularly for updates and send me bug reports and suggestions through GitHub: I will attempt to address those in future releases, time permitting.

Pre-requisites and Installation

The effective use of this JMtk15 package requires access to the following resources:

the Mtk shared library libMisrToolkit.so.1.5.0,

a set of third party libraries supporting HDF and NetCDF file formats, as well as additional utilities, and

MISR data files, as appropriate.

The Mtk and third party libraries can be installed on Intel x86_32 or x86_64 platforms as explained on the MISR Toolkit web site mentioned above. As of March 2023, the simplest approach to use this JMtk15 package on an Apple aarch_64 macBook equipped with a M1 or M2 chip is to port an existing working installation from an Intel-based platform and run under the Rosetta 2 environment: All that is required is to copy the Mtk and /usr/local/hdfeoslibs directories in exactly the same locations on the Mx platform, and to use the x86_64 version of Julia (see the juliaup for more details on installing multiple versions of Julia). Platform-specific libraries for the aarch_64 architecture may be developed at a later stage.

MISR data are accessible from the ASDC web site mentioned above.

Once those libraries and data files have been installed, edit the main module file JMtk15.jl to point to the Mtk library and the data directory. Look for the Package configuration section near the top of that file.

Index of functions

JMtk15.JMtk15

JMtk15.jMtkDateTimeToJulian

JMtk15.jMtkErrorMessage

JMtk15.jMtkFileAttrList

JMtk15.jMtkFileLGID

JMtk15.jMtkFileType

JMtk15.jMtkFileVersion

JMtk15.jMtkJulianToDateTime

JMtk15.jMtkVersion

Main module

JMtk15.JMtk15 — Module

using JMtk15

Purpose:

Provide Julia wrapper functions to the NASA JPL MISR Toolkit C functions.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia wrapper: Version 0.1.0 (2023-02-15).

Notes:

This module allows one to use some of the C functions defined in the MISR Toolkit (Mtk) as if they were Julia functions. Please read the README.md file and the documentation for this package to determine which functions are available and how they can be used.

The MISR Toolkit (Mtk) itself must be installed on your computer before this JMtk15 package can be used. In particular, the location of the shared library libMisrToolkit.so in the file system must be specified as part of the configuration of this package.

MISR data must be available on your computer, and their physical location must be specified as part of the configuration of this package.

This module must be imported before using any of the wrapper functions because it loads essential definitions and defines the location of resources (source files, test functions, documentation, data).

Example:

julia> using JMtk15
[ Info: Precompiling JMtk15 [6c71635d-e68b-418f-880a-c8128ca9dc0a]

Util

JMtk15.jMtkVersion — Function

version = jMtkVersion()

Purpose:

Report the version identifier of the MISR Toolkit (Mtk) library currently in use.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia wrapper: Version 0.1.0 (2023-02-15).

Example:

julia> using JMtk15

julia> version = jMtkVersion()
"1.5.0"

JMtk15.jMtkErrorMessage — Function

errmsg = jMtkErrorMessage(status)

Purpose:

Report the error message corresponding to the specified status code.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia wrapper: Version 0.1.0 (2023-02-15).

Examples:

julia> using JMtk15

julia> status = 0
0

julia> errmsg = jMtkErrorMessage(status)
"SUCCESS"

julia> status = 3
3

julia> errmsg = jMtkErrorMessage(status)
"BAD ARGUMENT"

JMtk15.jMtkDateTimeToJulian — Function

juldate = jMtkDateTimeToJulian(datetime)

Purpose:

Convert a date and time (in ISO 8601 format) into a Julian date.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia wrapper: Version 0.1.0 (2023-02-15).

Example:

julia> using JMtk15

julia> datetime = "2002-05-02T02:00:00Z"
"2002-05-02T02:00:00Z"

julia> juldate = jMtkDateTimeToJulian(datetime)
2.4523965833333335e6

JMtk15.jMtkJulianToDateTime — Function

datetime = jMtkJulianToDateTime(juldate)

Purpose:

Convert a Julian date into date and time (in ISO 8601 format).

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia wrapper: Version 0.1.0 (2023-02-15).

Example:

julia> using JMtk15

julia> juldate = 2.4523965833333335e6
2.4523965833333335e6

julia> datetime = jMtkJulianToDateTime(juldate)
"2002-05-02T02:00:00Z"

FileQuery

JMtk15.jMtkFileType — Function

filetype = jMtkFileType(filename)

Purpose:

Determine the MISR product file type.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia wrapper: Version 0.1.0 (2023-02-15).

Example:

julia> using JMtk15

julia> filename = JMtk15_data * "MISR/MISR_AM1_AS_LAND_P168_O068050_F08_0023.nc"
"/Users/michel/Codes/Julia/JMtk15/data/MISR/MISR_AM1_AS_LAND_P168_O068050_F08_0023.nc"

julia> filetype = jMtkFileType(filename)
"AS_LAND"

JMtk15.jMtkFileLGID — Function

lgid = jMtkFileLGID(filename)

Purpose:

Return the local granule ID of a MISR product file.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia wrapper: Version 0.1.0 (2023-02-15).

Example:

julia> using JMtk15

julia> filename = JMtk15_data * "MISR/MISR_AM1_GRP_ELLIPSOID_GM_P168_O068050_BF_F03_0024.hdf"
"/Users/michel/Codes/Julia/JMtk15/data/MISR/MISR_AM1_GRP_ELLIPSOID_GM_P168_O068050_BF_F03_0024.hdf"

julia> lgid = jMtkFileLGID(filename)
"MISR_AM1_GRP_ELLIPSOID_GM_P168_O068050_BF_F03_0024.hdf"

JMtk15.jMtkFileVersion — Function

fversion = jMtkFileVersion(filename)

Purpose:

Return the file version identifier of a MISR product file.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia wrapper: Version 0.1.0 (2023-02-15).

Example:

```julia julia> using JMtk15

julia> filename = JMtk15data * "MISR/MISRAM1GRPELLIPSOIDGMP168O068050BFF030024.hdf" "/Users/michel/Codes/Julia/JMtk15/data/MISR/MISRAM1GRPELLIPSOIDGMP168O068050BFF03_0024.hdf"

julia> fversion = jMtkFileVersion(filename) "F03_0024" ````

JMtk15.jMtkFileAttrList — Function

attrcnt, attrlist = jMtkFileAttrList(filename);

Purpose:

List the attributes of a specified MISR product file.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia wrapper: Version 0.1.0 (2023-02-15).

Example:

julia> using JMtk15

julia> filename = JMtk15_data * "MISR/MISR_AM1_GRP_ELLIPSOID_GM_P168_O068050_BF_F03_0024.hdf"
"/Users/michel/Codes/Julia/JMtk15/data/MISR/MISR_AM1_GRP_ELLIPSOID_GM_P168_O068050_BF_F03_0024.hdf"

julia> attrcnt, attrlist = jMtkFileAttrList(filename);

julia> attrcnt
31

julia> attrlist
31-element Vector{String}:
 "HDFEOSVersion"
 "StructMetadata.0"
 "Path_number"
 "AGP_version_id"
 "DID_version_id"
 "Number_blocks"
 "Ocean_blocks_size"
 "Ocean_blocks.count"
 "Ocean_blocks.numbers"
 "SOM_parameters.som_ellipsoid.a"
 "SOM_parameters.som_ellipsoid.e2"
 "SOM_parameters.som_orbit.aprime"
 "SOM_parameters.som_orbit.eprime"
 "SOM_parameters.som_orbit.gama"
 "SOM_parameters.som_orbit.nrev"
 ⋮
 "SOM_parameters.som_orbit.P2P1"
 "SOM_parameters.som_orbit.lambda0"
 "Origin_block.ulc.x"
 "Origin_block.ulc.y"
 "Origin_block.lrc.x"
 "Origin_block.lrc.y"
 "Start_block"
 "End block"
 "Cam_mode"
 "Num_local_modes"
 "Local_mode_site_name"
 "Orbit_QA"
 "Camera"
 "coremetadata"